// src/types/blog.types.ts
export type MediumFeed = {
  url: string;
  title: string;
  link: string;
  author: string;
  description: string;
  image: string;
};

export type MediumBlogItem = {
  title: string;
  pubDate: string;
  link: string;
  guid: string;
  author: string;
  thumbnail: string;
  description: string;
  content: string;
  categories: string[];
};

export type MediumApiResponse = {
  status: string;
  feed: MediumFeed;
  items: MediumBlogItem[];
};